dveri/requirements.txt
asgiref==3.7.2
Django==4.2.8
django-ckeditor==6.7.0
django-filter==23.5
django-js-asset==2.1.0
djangorestframework==3.14.0
importlib-metadata==7.0.0
Markdown==3.5.1
Pillow==10.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
